Inicio da atualização do keycloak
