+++ src/API/appsettings.json
@@ -13,5 +13,5 @@
   "ConnectionStrings": {
     "Base": "Server=db;Database=funcionarios;Port=5432;Username=postgres;Password=postgres;SSLMode=Prefer"
   },
-  "rabbit": "amqp://mc:mc2@192.168.0.143:5672/main"
+  "rabbit": "amqp://mc:mc2@192.168.15.32:5672/main"
 }
