+++ src/MVC/Cadastro.MVC.csproj
@@ -23,4 +23,7 @@
     </PackageReference>
     <PackageReference Include="Microsoft.AspNetCore.Mvc.NewtonsoftJson" Version="6.0.2" />
   </ItemGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\Cadastro.Configuracoes\Cadastro.Configuracoes.csproj" />
+  </ItemGroup>
 </Project>
