--- src/MVC/Configuration/HasScopeRequirement.cs
@@ -1,17 +1,0 @@
-﻿using Microsoft.AspNetCore.Authorization;
-using System;
-
-namespace API.Configuration
-{
-    public class HasScopeRequirement : IAuthorizationRequirement
-    {
-        public string Issuer { get; }
-        public string Scope { get; }
-
-        public HasScopeRequirement(string scope, string issuer)
-        {
-            Scope = scope ?? throw new ArgumentNullException(nameof(scope));
-            Issuer = issuer ?? throw new ArgumentNullException(nameof(issuer));
-        }
-    }
-}
