+++ src/Rabbitmq/Dockerfile
@@ -1,4 +1,4 @@
-FROM rabbitmq:3.9-management-alpine
+FROM rabbitmq:3.9.13-management-alpine
 
 RUN rabbitmq-plugins enable --offline rabbitmq_management
 RUN rabbitmq-plugins enable --offline rabbitmq_shovel rabbitmq_shovel_management
